chore (api-cat) Move ApiRegistrationPublic class to api-cat-common, so that it could share common types with ApiDocument
