+++ applications/api-cat/pom.xml
@@ -99,12 +99,12 @@
         <dependency>
             <groupId>no.dcat</groupId>
             <artifactId>api-cat-bindings</artifactId>
-            <version>1.5.0</version>
+            <version>1.6.0</version>
         </dependency>
         <dependency>
             <groupId>no.dcat</groupId>
             <artifactId>api-cat-common</artifactId>
-            <version>1.0.0</version>
+            <version>1.1.0</version>
         </dependency>
         <dependency>
             <groupId>no.dcat</groupId>
@@ -114,7 +114,7 @@
         <dependency>
             <groupId>no.dcat</groupId>
             <artifactId>registration-api-bindings</artifactId>
-            <version>1.1.0</version>
+            <version>1.2.0</version>
         </dependency>
         <dependency>
             <groupId>no.dcat</groupId>
