+++ applications/api-cat/src/main/java/no/acat/service/ApiDocumentBuilderService.java
@@ -5,11 +5,11 @@ import no.acat.repository.ApiDocumentRepository;
 import no.dcat.client.publishercat.PublisherCatClient;
 import no.dcat.htmlclean.HtmlCleaner;
 import no.dcat.shared.*;
+import no.fdk.acat.common.model.ApiRegistrationPublic;
 import no.fdk.acat.common.model.apispecification.ApiSpecification;
 import no.fdk.acat.common.model.apispecification.ExternalDocumentation;
 import no.fdk.acat.converters.apispecificationparser.ParseException;
 import no.fdk.acat.converters.apispecificationparser.UniversalParser;
-import no.fdk.registration.common.ApiRegistrationPublic;
 import org.apache.commons.io.IOUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
