+++ applications/api-cat/src/test/java/no/acat/service/ApiDocumentBuilderServiceTest.java
@@ -5,7 +5,7 @@ import no.acat.repository.ApiDocumentRepository;
 import no.acat.utils.Utils;
 import no.fdk.test.testcategories.UnitTest;
 import no.fdk.acat.converters.apispecificationparser.ParseException;
-import no.fdk.registration.common.ApiRegistrationPublic;
+import no.fdk.acat.common.model.ApiRegistrationPublic;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
