+++ applications/informationmodel-cat/pom.xml
@@ -90,7 +90,7 @@
         <dependency>
             <groupId>no.dcat</groupId>
             <artifactId>registration-api-bindings</artifactId>
-            <version>1.1.0</version>
+            <version>1.2.0</version>
         </dependency>
         <dependency>
             <groupId>no.dcat</groupId>
