+++ libraries/api-cat-common/pom.xml
@@ -7,7 +7,7 @@
 
     <groupId>no.dcat</groupId>
     <artifactId>api-cat-common</artifactId>
-    <version>1.0.0</version>
+    <version>1.1.0</version>
 
     <properties>
         <!--standard properties-->
@@ -29,6 +29,11 @@
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-annotations</artifactId>
             <version>2.9.8</version>
+        </dependency>
+        <dependency>
+            <groupId>io.swagger</groupId>
+            <artifactId>swagger-annotations</artifactId>
+            <version>1.5.21</version>
         </dependency>
     </dependencies>
 </project>
