+++ libraries/api-cat-converters/pom.xml
@@ -59,7 +59,7 @@
         <dependency>
             <groupId>no.dcat</groupId>
             <artifactId>api-cat-common</artifactId>
-            <version>1.0.0</version>
+            <version>1.1.0</version>
             <scope>compile</scope>
         </dependency>
         <dependency>
