+++ libraries/informationmodel-cat-bindings/pom.xml
@@ -14,7 +14,7 @@
 
     <groupId>no.dcat</groupId>
     <artifactId>informationmodel-cat-bindings</artifactId>
-    <version>1.1.0</version>
+    <version>1.2.0</version>
 
     <dependencies>
         <dependency>
@@ -24,8 +24,7 @@
         <dependency>
             <groupId>no.dcat</groupId>
             <artifactId>api-cat-common</artifactId>
-            <version>1.0.0</version>
-            <scope>compile</scope>
+            <version>1.1.0</version>
         </dependency>
         <dependency>
             <groupId>org.slf4j</groupId>
