+++ libraries/pom.xml
@@ -25,7 +25,6 @@
         <module>publisher-cat-client</module>
         <module>referencedata-client</module>
         <module>registration-api-bindings</module>
-        <module>registration-api-common</module>
         <module>shared</module>
         <module>test</module>
         <module>webutils-spring</module>
