+++ libraries/registration-api-bindings/pom.xml
@@ -14,7 +14,7 @@
 
     <groupId>no.dcat</groupId>
     <artifactId>registration-api-bindings</artifactId>
-    <version>1.1.0</version>
+    <version>1.2.0</version>
     <properties>
         <!--standard properties-->
         <java.version>1.8</java.version>
@@ -36,9 +36,10 @@
         </dependency>
         <dependency>
             <groupId>no.dcat</groupId>
-            <artifactId>registration-api-common</artifactId>
+            <artifactId>api-cat-common</artifactId>
             <version>1.1.0</version>
         </dependency>
+
 
     </dependencies>
 </project>
